+++ src/files.ts
@@ -1,20 +1,22 @@
-export class WSFileRef {
-	private id: number;
-	private path: string;
-	private words: number;
-	private lastUpdate: number;
-	private title: string;
-	private tags: string[];
-	private links: Map<WSFileRef, string>;
+export class WSFile {
+	name: string;
+	path: string;
+	private currentWords: number;
+	lastUpdate: number;
+	private ftitle: string;
+	tags: string[];
+	links: Map<WSFile, string>;
+	backlinks: WSFile[];
 
-	constructor(id: number, path: string) {
-		this.id = id;
+	constructor(name: string, path: string) {
+		this.name = name;
 		this.path = path;
-		this.words = 0;
+		this.currentWords = 0;
 		this.lastUpdate = Date.now();
-		this.title = null;
+		this.ftitle = null;
 		this.tags = [];
-		this.links = new Map<WSFileRef, string>();
+		this.links = new Map<WSFile, string>();
+		this.backlinks = [];
 	}
 
 	clearTags() {
@@ -33,35 +35,52 @@ export class WSFileRef {
 		this.tags = tags;
 	}
 
-	getTags() {
-		return this.tags;
-	}
-
 	hasTag(tag: string) {
 		return this.tags.contains(tag);
 	}
 
 	clearLinks() {
+		this.links.forEach((name:string, file:WSFile) => {
+			file.removeBacklink(this);
+		})
 		this.links.clear();
 	}
 
-	setLink(file: WSFileRef, title: string) {
+	setLink(file: WSFile, title: string) {
 		this.links.set(file, title);
+		file.setBacklink(this);
 	}
 
-	hasLink(file: WSFileRef) {
+	setBacklink(file: WSFile) {
+		if (!this.backlinks.contains(file)) {
+			this.backlinks.push(this);
+		}
+	}
+
+	hasBacklink(file: WSFile) {
+		return (this.backlinks.contains(file));
+	}
+
+	removeBacklink(file: WSFile) {
+		if (this.backlinks.contains(file)) {
+			this.backlinks.remove(file);
+		}
+	}
+
+	hasLink(file: WSFile) {
 		return this.links.has(file);
 	}
 
-	removeLink(file: WSFileRef) {
+	removeLink(file: WSFile) {
 		if (this.links.has(file)) {
 			this.links.delete(file);
+			file.removeBacklink(this);
 		} else {
 			console.log("Tried to remove link to '%s' from WSFileRef(%s), but it is not there.", file.path, this.path);
 		}
 	}
 
-	getLinkTitle(file: WSFileRef) {
+	getLinkTitle(file: WSFile) {
 		return this.links.get(file);
 	}
 
@@ -69,28 +88,25 @@ export class WSFileRef {
 		return Array.from(this.links.keys());
 	}
 
-	getLinks() {
-		return this.links;
-	}
-
 	setTitle(newTitle: string) {
 		if (newTitle == "" || newTitle == undefined || newTitle == null) {
-			this.title = null;
+			this.ftitle = null;
 		} else {
-			this.title = newTitle;
+			this.ftitle = newTitle;
 		}
 	}
 
-	getTitle() {
-		if (this.title == "" || this.title == null) {
-			let title = this.path.split("/").pop().replace(/\.md$/, "");
-			return title;
+	get title() {
+		if (this.ftitle == "" || this.ftitle == null) {
+			//let title = this.path.split("/").pop().replace(/\.md$/, "");
+			//return title;
+			return this.name;
 		}
-		return this.title;
+		return this.ftitle;
 	}
 
 	hasTitle() {
-		return this.title != null;
+		return this.ftitle != null;
 	}
 
 	setPath(newPath: string) {
@@ -98,29 +114,17 @@ export class WSFileRef {
 		this.lastUpdate = Date.now();
 	}
 
-	getPath(): string {
-		return this.path;
-	}
-
 	setWords(count: number) {
-		this.words = count;
+		this.currentWords = count;
 		this.lastUpdate = Date.now();
 	}
 
 	addWords(count: number) {
-		this.words += count;
+		this.currentWords += count;
 		this.lastUpdate = Date.now();
 	}
 
-	getWords(): number {
-		return this.words;
-	}
-
-	getID(): number {
-		return this.id;
-	}
-
-	getLastUpdate(): number {
-		return this.lastUpdate;
+	get words() {
+		return this.currentWords;
 	}
 }