+++ src/projects.ts
@@ -1,293 +1,315 @@
-import { TFile } from "obsidian";
 import { WSDataCollector } from "./data";
-import { WSFileRef } from "./files";
+import { WSFile } from "./files";
 import WordStatisticsPlugin from "./main";
 
-interface WSProjectData {
-    displayText: string,
-    wordsExcluded: boolean,
+export enum WSPType {
+    File = 0,
+    Folder,
+    Tag
 }
 
-enum WSPIndexType {
-    File = 0, Folder, Tag
-}
+const PROJECT_TYPE_STRING = ["WSFileProject", "WSFolderProject", "WSTagProject"];
 
-export interface WSProjectMap {
-    name: string;
-    type: WSPIndexType;
+/*
+
+interface Project {
+    name: string,
+    readonly type: WSPType,
     index: string;
 }
 
-export class WSProjectManager {
-    private plugin: WordStatisticsPlugin;
-    private collector: WSDataCollector;
-    private projects: Map<string, WSProject>;
-    private indexFiles: Map<WSFileRef, WSProject>;
+interface Group {
+    name: string,
+    projects: string[];
+}
 
-    constructor(plugin: WordStatisticsPlugin, collector: WSDataCollector) {
-        this.plugin = plugin;
-        this.collector = collector;
-        this.projects = new Map<string, WSProject>();
-        this.indexFiles = new Map<WSFileRef, WSProject>();
-    }
+interface ProjectMap {
+    fileProjects: Project[];
+    folderProjects: Project[];
+    tagProjects: Project[];
+    projectGroups: Group[];
+}
 
-    async buildProjects(projects: WSProjectMap[]) {
-        projects.forEach((project: WSProjectMap) => {
-            if (this.newProject(project) == null) {
-                console.log("Unable to create project for ", project);
-                throw (Error());
-            }
-        });
-    }
+class WSProject {
+    name: string;
+    readonly type: WSPType;
 
-    mapProjects() {
-        let maps: WSProjectMap[] = [];
-        for (const [name, project] of this.projects) {
-            let map: WSProjectMap = {
-                name: project.getName(),
-                type: project.getIndexType(),
-                index: project.getIndexString()
-            };
-            maps.push(map);
-        }
-        return maps;
-    }
+    constructor(name: string, type: WSPType) {
+        this.name = name;
+        this.type = type;
+    };
 
-    isIndexFile(fi: WSFileRef) {
-        return this.indexFiles.has(fi);
-    }
-
-    projectNameExists(name: string) {
-        return this.projects.has(name);
-    }
-
-    // unless we're using some kind of ID system, there's no real way for this to be useful
-    // if you change the name of the project, then it will no longer trigger in all of the links
-    // and there will instead be project files linked to an invalid project, which I suppose
-    // could then be listed in the project leaf;
-    // we don't want the project to be manipulating any files, so changing the YAML tags for a file
-    // doesn't seem like a good plan
-    renameProject(project: WSProject, name: string) {
-        let oldName = project.getName();
-        project.setName(name);
-        this.projects.delete(oldName);
-        this.projects.set(name, project);
-    }
-
-    getProject(name: string) {
-        if (!this.projects.has(name)) {
-            return null;
-        }
-        return this.projects.get(name);
-    }
-
-    newProject(map: WSProjectMap) {
-        if (this.projects.has(map.name)) {
-            return null;
-        }
-        let project = new WSProject(this, this.collector, map);
-        if (project.indexedByFile()) {
-            this.indexFiles.set(project.getIndexFile(), project);
-        }
-        this.projects.set(map.name, project);
-        return project;
-    }
-
-    updateProject(name: string, map: WSProjectMap) {
-        let project = this.projects.get(name);
-        let oldMap: WSProjectMap = {
-            name: project.getName(),
-            type: project.getIndexType(),
-            index: project.getIndexString()
-        };
-
-        if (oldMap.type === WSPIndexType.File && map.type != WSPIndexType.File) {
-            this.indexFiles.delete(project.getIndexFile());
-        }
-        let rebuild = project.update(map);
-        if (oldMap.name != map.name) {
-            this.renameProject(project, map.name);
-        }
-        if (rebuild) {
-            project.buildIndex();
-        }
-    }
-
-    deleteProject(name: string) {
-        if (!this.projects.has(name)) {
-            console.log("Tried to delete project '%s' but it does not exist.", name);
-        } else {
-            this.projects.get(name).cleanup();
-            this.projects.delete(name);
-        }
-    }
-
-    getProjectList() {
-        return (Array.from(this.projects.values()));
-    }
-
-    getProjectNames() {
-        return (Array.from(this.projects.keys()));
-    }
-
-    getProjectsCount() {
-        return this.projects.size;
-    }
-
-    cleanup() {
-        for (const [, project] of this.projects) {
-            project.cleanup();
-        }
-        this.projects.clear();
+    serialize() {
+        return { name: this.name, type: this.type, index: "" };
     }
 }
 
-export class WSProject {
-    private name: string;
-    private manager: WSProjectManager;
-    private collector: WSDataCollector;
-    private files: Map<WSFileRef, WSProjectData>;
-    private indexType: WSPIndexType;
-    private indexFolder: string; // should we find a way to encapsulate folders?
-    private indexFile: WSFileRef;
-    private indexTag: string;
+export class WSFileProject extends WSProject {
+    file: WSFile;
 
-    constructor(manager: WSProjectManager, collector: WSDataCollector, map: WSProjectMap) {
-        this.manager = manager;
-        this.collector = collector;
-        this.name = map.name;
-        this.indexType = map.type;
-        this.indexFolder = null;
-        this.indexFile = null;
-        this.indexTag = null;
-        this.files = new Map<WSFileRef, WSProjectData>();
-        if (this.update(map)) {
-            this.buildIndex();
-        }
+    constructor(name: string, file: WSFile) {
+        super(name, WSPType.File);
+        this.file = file;
     }
 
-    buildIndex() {
-        this.cleanup();
-        if (this.indexedByFile()) {
-            let fileRef = this.indexFile;
-            if (fileRef === null) {
-                console.log("Invalid index file ('%s') for project ('%s')", this.indexFile, this.name, this);
-                throw (Error("Invalid file index passed for project."));
-            }
+    override serialize() {
+        return { name: this.name, type: this.type, index: this.file.path };
+    }
+}
 
-        }
+export class WSFolderProject extends WSProject {
+    folder: string;
+
+    constructor(name: string, folder: string) {
+        super(name, WSPType.Folder);
+        this.folder = folder;
     }
 
-    update(map: WSProjectMap) {
-        let rebuild = (this.indexType != map.type) || (this.indexType === map.type && this.getIndexString() != map.index);
-        if (map.type === WSPIndexType.File) {
-            this.indexFile = this.collector.getFile(map.index);
-        } else if (map.type === WSPIndexType.Folder) {
-            this.indexFolder = map.index;
-        } else if (map.type === WSPIndexType.Tag) {
-            this.indexTag = map.index;
-        }
-        return rebuild;
+    override serialize() {
+        return { name: this.name, type: this.type, index: this.folder };
+    }
+}
+
+export class WSTagProject extends WSProject {
+    tag: string;
+    override readonly type: WSPType = WSPType.Tag;
+
+    constructor(name: string, tag: string) {
+        super(name, WSPType.Tag);
+        this.tag = tag;
     }
 
-    indexedByFolder() {
-        return this.indexType === WSPIndexType.Folder;
+    override serialize() {
+        return { name: this.name, type: this.type, index: this.tag };
     }
+}
 
-    indexedByFile() {
-        return this.indexType === WSPIndexType.File;
-    }
 
-    indexedByTag() {
-        return this.indexType === WSPIndexType.Tag;
-    }
+export class WSProjectGroup {
+    name: string;
+    projects: WSProject[];
 
-    setIndexFile(file: WSFileRef) {
-        this.indexType = WSPIndexType.File;
-        this.indexFile = file;
-        this.indexFolder = null;
-        this.indexTag = null;
-    }
-
-    setIndexFolder(path: string) {
-        this.indexType = WSPIndexType.Folder;
-        this.indexFile = null;
-        this.indexFolder = path;
-        this.indexTag = null;
-    }
-
-    setIndexTag(tag: string) {
-        this.indexType = WSPIndexType.Tag;
-        this.indexFile = null;
-        this.indexFolder = null;
-        this.indexTag = tag;
-    }
-
-    getIndexFile() {
-        return this.indexType === WSPIndexType.File ? this.indexFile : null;
-    }
-
-    getIndexFolder() {
-        return this.indexType === WSPIndexType.Folder ? this.indexFolder : null;
-    }
-
-    getIndexTag() {
-        return this.indexType === WSPIndexType.Tag ? this.indexTag : null;
-    }
-
-    getIndexString() {
-        if (this.indexedByFile()) {
-            return (this.indexFile != null && this.indexFile != undefined) ? this.indexFile.getPath() : null;
-        } else if (this.indexedByFolder()) {
-            return this.indexFolder;
-        } else if (this.indexedByTag()) {
-            return this.indexTag;
-        }
-    }
-
-    getIndexType() {
-        return this.indexType;
-    }
-
-    cleanup() {
-        this.files.clear();
-    }
-
-    getName() {
-        return this.name;
-    }
-
-    setName(name: string) {
+    constructor(name: string, projects: WSProject[]) {
         this.name = name;
+        this.projects = projects;
+    };
+
+    serialize() {
+        let projList: string[] = [];
+        this.projects.forEach((proj: WSProject) => {
+            projList.push(proj.name);
+        });
+        return { name: this.name, projects: projList };
+    }
+}
+
+*/
+
+interface WSProject {
+    name: string,
+    readonly type: WSPType,
+    index: string;
+}
+
+interface WSProjectGroup {
+    name: string,
+    projects: string[];
+}
+
+interface WSProjectMap {
+    fileProjects: WSProject[];
+    folderProjects: WSProject[];
+    tagProjects: WSProject[];
+    projectGroups: Map<string, WSProjectGroup>;
+}
+
+
+/*
+
+A Longform project is basically a project group, but it's a special case. Keeping track of its history may be somewhat boggling, though.
+I can always auto-generate them on startup, but it makes more sense to generate them once and then just keep track of them.
+
+People who are using Longform are likely to use Obsidian for all of the writing process and aren't going to change files outside, hopefully.
+
+*/
+
+export class WSProjectManager {
+    tagProjects: WSProject[] = [];
+    fileProjects: WSProject[] = [];
+    folderProjects: WSProject[] = [];
+    projectGroups: Map<string, WSProjectGroup>;
+    plugin: WordStatisticsPlugin = null;
+    collector: WSDataCollector = null;
+    projects: Map<string, [WSPType, WSProject]>;
+    errorState: boolean = false;
+    errorMessages: string[] = [];
+
+    constructor(plugin: WordStatisticsPlugin, collector: WSDataCollector, wMap?: WSProjectMap) {
+        this.plugin = plugin;
+        this.collector = collector;
+        this.projects = new Map<string, [WSPType, WSProject]>();
+        this.projectGroups = new Map<string, WSProjectGroup>();
+        if (wMap != undefined) {
+            this.loadProjects(wMap.fileProjects);
+            this.loadProjects(wMap.folderProjects);
+            this.loadProjects(wMap.tagProjects);
+            wMap.projectGroups.forEach((group) => {
+                this.realizeProjectGroup(group);
+            });
+        }
     }
 
-    addFile(file: WSFileRef, name: string, excluded: boolean) {
-        if (!this.files.has(file)) {
-            this.files.set(file, { displayText: name, wordsExcluded: excluded });
+    checkProjectName(name: string) {
+        return !this.projects.has(name);
+    }
+
+    getProject(name: string) {
+        let [, proj] = this.projects.get(name);
+        return proj;
+    }
+
+    getProjectsForGroup(name: string) {
+        if (this.projectGroups.has(name)) {
+            let projects: WSProject[] = [];
+            this.projectGroups.get(name).projects.forEach((pName) => {
+                let [, proj] = this.projects.get(name);
+                if (proj != undefined && proj != null) {
+                    projects.push(proj);
+                }
+            });
+            return projects;
+        }
+        return undefined;
+    }
+
+    getIndexFile(proj: WSProject) {
+        if (proj.type === WSPType.File) {
+            return this.collector.getFile(proj.index);
+        }
+        return undefined;
+    }
+
+    getIndexTag(proj: WSProject) {
+        if (proj.type === WSPType.Tag) {
+            return proj.index;
+        }
+        return undefined;
+    }
+
+    getIndexFolder(proj: WSProject) {
+        if (proj.type === WSPType.Folder) {
+            return proj.index;
+        }
+        return undefined;
+    }
+
+    logError(msg: string) {
+        this.errorState = true;
+        this.errorMessages.push(msg);
+        console.log("WordStatistics.ProjectManager[err]: " + msg);
+    }
+
+    registerProject(proj: WSProject) {
+        if (this.projects.has(proj.name)) {
+            console.log(this.projects, proj, proj.name);
+            this.logError(`Tried to register project '${proj.name}', but one with that name already exists of type '${PROJECT_TYPE_STRING[proj.type]}'.`);
+            throw Error();
+        }
+        this.projects.set(proj.name, [proj.type, proj]);
+        switch (proj.type) {
+            case WSPType.File:
+                this.fileProjects.push(proj);
+                break;
+            case WSPType.Folder:
+                this.folderProjects.push(proj);
+                break;
+            case WSPType.Tag:
+                this.tagProjects.push(proj);
+                break;
+            default:
+                console.log(proj);
+                this.logError(`Invalid project type: ${proj.type} for project: ${proj}`);
+        }
+    }
+    /*
+        realizeProject(proj: Project) {
+            if (proj.type === WSPType.File) {
+                return (new WSFileProject(proj.name, this.collector.getFile(proj.index)));
+            } else if (proj.type === WSPType.Folder) {
+                return (new WSFolderProject(proj.name, proj.index));
+            } else if (proj.type === WSPType.Tag) {
+                return (new WSTagProject(proj.name, proj.index));
+            } else {
+                throw Error(`Invalid project type ${proj.type}`);
+            }
+        }
+    */
+    validateProjectLoad(projects: WSProject[]) {
+        let names: string[] = [];
+        projects.forEach((proj) => {
+            if (names.contains(proj.name)) {
+                console.log(`Project name already in use: ${proj.name}`);
+                return false;
+            }
+            names.push(proj.name);
+        });
+        return true;
+    }
+
+    loadProjects(projects: WSProject[]) {
+        if (this.validateProjectLoad(projects)) {
+            projects.forEach((proj: WSProject) => {
+                this.registerProject(proj);
+            });
+        }
+    }
+
+    realizeProjectGroup(group: WSProjectGroup) {
+        if (this.projectGroups.has(group.name)) {
+            console.log(this.projectGroups.get(group.name));
+            this.logError(`Tried to realize project group '${group.name}' but a group already exists with that name.`);
+            return;
+        }
+
+        let fail = false;
+        group.projects.forEach((name: string) => {
+            if (!this.projects.has(name)) {
+                fail = true;
+                this.logError(`Could not resolve project '${name}' for project group '${group.name}'.`);
+            }
+        });
+        if (fail) {
+            this.logError(`Failed to load project group '${group.name}'.`);
         } else {
-            console.log("Tried to add WSFileRef(%d)[%s] to project '%s', but it was already in file list.", file.getID(), file.getPath(), this.name);
+            this.projectGroups.set(group.name, group);
         }
     }
 
-    hasFile(file: WSFileRef) {
-        return (this.files.has(file));
-    }
-
-    deleteFile(file: WSFileRef) {
-        if (this.files.has(file)) {
-            this.files.delete(file);
-        } else {
-            console.log("Tried to remove WSFileRef(%d)[%s] from project '%s', but it was not there.", file.getID(), file.getPath(), this.name);
+    /*
+        serialize(): ProjectMap {
+            let fileProjects: Project[] = [];
+            let folderProjects: Project[] = [];
+            let tagProjects: Project[] = [];
+            let projectGroups: Group[] = [];
+    
+            this.fileProjects.forEach((proj) => {
+                fileProjects.push(proj.serialize());
+            });
+            this.folderProjects.forEach((proj) => {
+                folderProjects.push(proj.serialize());
+            });
+            this.tagProjects.forEach((proj) => {
+                tagProjects.push(proj.serialize());
+            });
+            this.projectGroups.forEach((grp) => {
+                projectGroups.push(grp.serialize());
+            });
+            return { fileProjects, folderProjects, tagProjects, projectGroups };
         }
-    }
+    */
 
-    getFilesTuple() {
-        let files: [WSFileRef, WSProjectData][] = [];
-        let fileList = Array.from(this.files.keys());
-        for (let i = 0; i < fileList.length; i++) {
-            files.push([fileList[i], this.files.get(fileList[i])]);
-        }
-        return files;
+    serialize(): WSProjectMap {
+        return { fileProjects: this.fileProjects, folderProjects: this.folderProjects, tagProjects: this.tagProjects, projectGroups: this.projectGroups };
     }
 
 }